chore: Move files and correct component path for shadcn
