+++ components.json
@@ -12,7 +12,7 @@
   },
   "iconLibrary": "lucide",
   "aliases": {
-    "components": "@/components",
+    "components": "@/features/shared/components",
     "utils": "@/lib/utils",
     "ui": "@/features/shared/components/ui",
     "lib": "@/lib",
